Move table name validation function to plugins.utils
